+++ CHANGELOG.md
@@ -8,6 +8,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 3.0.1 
+
+### Added 
+
+- More logging from `elm-dialog-proxy`
+
+### Changed 
+
+- `elm-dialog-proxy` now makes multiple attempts to subscribe to the
+  dialog element
+
 ## 3.0.0
 
 ### Added